// extensions/nodejs-engine/src/model/base-schema.ts
export interface components {
    schemas: {
// ...
        Event: {
            id?: string;
            /**
             * Format: enum
             * @enum {string}
             */
            action?: "CREATE" | "UPDATE" | "DELETE" | "GET" | "LIST" | "OPERATE";
            actionSummary?: string;
            actionDescription?: string;
            resource?: components["schemas"]["Resource"];
            /** @description CREATE,UPDATE - records are for incoming and outgoing records GET - there will be only one record LIST - result of the list operation */
            records?: (components["schemas"]["Record"])[];
            /** @description GET - there will be only one id, for getting record with id DELETE - there will be multiple ids, for deleting multiple records */
            ids?: (string)[];
            recordSearchParams?: components["schemas"]["Event_RecordSearchParams"];
            /** @description If true, this will be last event on operation list */
            finalizes?: boolean;
            /** @description If true, backend will wait for processing this event before sending next one on operation list */
            sync?: boolean;
            /** Format: date-time */
            time?: string;
            /** @description Request annotations */
            annotations?: {
                [key: string]: string | undefined;
            };
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}